main/requirements.txt
flask
pandas
transformers
konlpy
wordcloud
torch
scikit-learn
sentencepiece
fasttext-wheel
wheel